VMM/HM: Explicit cast to 64-bit while flipping X86_CR4_VMXE.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@80871 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/VMM/VMMR0/HMR0.cpp
@@ -447,7 +447,7 @@ static int hmR0InitIntel(void)
 
             /* Restore CR4.VMXE if it wasn't set prior to us setting it above. */
             if (!(uOldCr4 & X86_CR4_VMXE))
-                SUPR0ChangeCR4(0 /* fOrMask */, ~X86_CR4_VMXE);
+                SUPR0ChangeCR4(0 /* fOrMask */, ~(uint64_t)X86_CR4_VMXE);
 
             /* Restore interrupts. */
             ASMSetFlags(fEFlags);
