+++ src/VBox/VMM/VMMR0/HMVMXR0.cpp
@@ -1620,7 +1620,7 @@ static int hmR0VmxEnterRootMode(PVMCC pVM, RTHCPHYS HCPhysCpuPage, void *pvCpuPa
     {
         /* Restore CR4.VMXE if it was not set prior to our attempt to set it above. */
         if (!(uOldCr4 & X86_CR4_VMXE))
-            SUPR0ChangeCR4(0 /* fOrMask */, ~X86_CR4_VMXE);
+            SUPR0ChangeCR4(0 /* fOrMask */, ~(uint64_t)X86_CR4_VMXE);
 
         if (pVM)
             pVM->hm.s.vmx.HCPhysVmxEnableError = HCPhysCpuPage;
@@ -1652,7 +1652,7 @@ static int hmR0VmxLeaveRootMode(void)
     {
         /* Exit VMX root mode and clear the VMX bit in CR4. */
         VMXDisable();
-        SUPR0ChangeCR4(0 /* fOrMask */, ~X86_CR4_VMXE);
+        SUPR0ChangeCR4(0 /* fOrMask */, ~(uint64_t)X86_CR4_VMXE);
         rc = VINF_SUCCESS;
     }
     else
